move `pycroft.lib.user` contents into `_old` module
